requirements.txt
altgraph==0.17.3
bleach==6.0.0
build==0.10.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.1.0
docutils==0.20.1
exrex==0.11.0
google-api-core==2.12.0
google-auth==2.23.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.7
importlib-metadata==6.7.0
jaraco.classes==3.2.3
keyring==24.2.0
macholib==1.16.2
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
twine==4.0.2
urllib3==2.0.7
webencodings==0.5.1
zipp==3.15.0
